# docs/swagger.yaml
definitions:
  github_com_ruziba3vich_soand_internal_models.PinChatRequest:
    properties:
      chat_id:
        type: string
      pin:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Soand API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Soand API
  version: "1.0"
paths:
...
  /chats/pin:
    post:
      consumes:
      - application/json
      description: Pins or unpins a chat for the authenticated user
      parameters:
      - description: Pin chat request
        in: body
        name: pinChatRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_ruziba3vich_soand_internal_models.PinChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pin or unpin a chat
      tags:
      - chats
  /chats/pinned:
    get:
      description: Retrieves a paginated list of pinned chats for the authenticated
        user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of pinned chats
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pinned chats
      tags:
      - chats
...
  /posts:
    get:
      consumes:
      - application/json
      description: Retrieves a single post using its MongoDB ObjectID
      parameters:
      - description: Post ID (MongoDB ObjectID)
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post details
          schema: {}
        "400":
          description: Invalid post ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a post by ID
      tags:
      - posts
    post:
      consumes:
      - multipart/form-data
      description: Creates a post with description, tags, optional delete_after time,
        and file attachments
      parameters:
      - description: Post description
        in: formData
        name: description
        required: true
        type: string
      - description: Time in minutes after which the post will be deleted
        in: formData
        name: delete_after
        type: string
      - description: Comma-separated list of tags or JSON array
        in: formData
        name: tags
        type: string
      - description: JSON stringified array of tags (alternative to tags)
        in: formData
        name: tags_json
        type: string
      - description: Files to upload (multiple allowed)
        in: formData
        name: files
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Post created successfully with ID
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a post using its MongoDB ObjectID
      parameters:
      - description: Post ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid post ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete post
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Updates a post by ID with new data
      parameters:
      - description: Post ID (MongoDB ObjectID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (e.g., description, tags)
        in: body
        name: updateData
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid post ID or payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update post
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - posts
  /posts/all:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of all posts
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: string
      - description: 'Number of posts per page (default: 10)'
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of posts
          schema:
            items: {}
            type: array
        "500":
          description: Failed to retrieve posts
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all posts
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: '"JWT Authorization header using the Bearer scheme. Example: ''Bearer
      {token}''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"